WID-195 - fix, don't filter out line breaks if filter doesn't exist, added extra checks

+++ src/app/widget-builder/components/widgets/base-widget-edit.component.ts
@@ -29,7 +29,12 @@ export class BaseWidgetEditComponent extends AbstractWidgetEditComponent impleme
     // Subscribe to changes in the form and reflect them on the widget model
     this.formSubscription = this.widgetEditForm.valueChanges.subscribe(values => {
       // Remove Line breaks from search_query
-      values.general.search_query = this.queryStringService.removeLineBreaks(values.general.search_query);
+      if(values.search_params.query){
+        values.search_params.query = this.queryStringService.removeLineBreaks(values.search_params.query)
+      }
+      if(values.general.search_query){
+        values.general.search_query = this.queryStringService.removeLineBreaks(values.general.search_query);
+      }
       // Apply the values to the model
       this.applyValuesToModel(values);
     });
